Add helper class and test boilerplate
Must be considered a work in progress.

Added:
 - a helper class to handle the plugins that have a common interface
 - a test skeleton to check that the handle works

The build lacks automation for now. The required dlls to let the test work need to be copied where the test executable resides.
At the current state, the goal is to find an appropriate way to pass the path to the plugin dll to the test executable.